fix(onboarding): pages refactoring
