+++ libs/pages/login/feature/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/login-page'
+export * from './lib/login'
