--- libs/pages/login/feature/src/lib/login-page.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react'
-
-import PagesLoginFeature from './login-page'
-
-describe('PagesLoginFeature', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<PagesLoginFeature />)
-    expect(baseElement).toBeTruthy()
-  })
-})
