+++ libs/pages/login/feature/src/lib/login.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react'
+
+import Login from './login'
+
+describe('Login', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<Login />)
+    expect(baseElement).toBeTruthy()
+  })
+})
