+++ libs/pages/onboarding/feature/src/index.ts
@@ -1,6 +1,6 @@
-export * from './lib/pages/onboarding-project-page/onboarding-project-page'
-export * from './lib/pages/onboarding-pricing-page/onboarding-pricing-page'
-export * from './lib/pages/onboarding-more-page/onboarding-more-page'
-export * from './lib/pages/onboarding-company-page/onboarding-company-page'
-export * from './lib/pages/onboarding-personalize-page/onboarding-personalize-page'
-export * from './lib/onboarding-page'
+export * from './lib/components/onboarding-project/onboarding-project'
+export * from './lib/components/onboarding-pricing/onboarding-pricing'
+export * from './lib/components/onboarding-more/onboarding-more'
+export * from './lib/components/onboarding-company/onboarding-company'
+export * from './lib/components/onboarding-personalize/onboarding-personalize'
+export * from './lib/onboarding'
