+++ libs/pages/onboarding/feature/src/lib/components/container/container.spec.tsx
@@ -1,0 +1,9 @@
+import { render } from '@testing-library/react'
+import Container from './container'
+
+describe('Container', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<Container />)
+    expect(baseElement).toBeTruthy()
+  })
+})
