+++ libs/pages/onboarding/feature/src/lib/components/onboarding-company/onboarding-company-page.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react'
+
+import OnboardingCompany from './onboarding-company'
+
+describe('OnboardingCompany', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<OnboardingCompany />)
+    expect(baseElement).toBeTruthy()
+  })
+})
