+++ libs/pages/onboarding/feature/src/lib/components/onboarding-more/onboarding-more.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react'
+
+import OnboardingMore from './onboarding-more'
+
+describe('OnboardingMore', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<OnboardingMore />)
+    expect(baseElement).toBeTruthy()
+  })
+})
