+++ libs/pages/onboarding/feature/src/lib/components/onboarding-more/onboarding-more.tsx
@@ -1,7 +1,7 @@
 import { StepMore } from '@console/pages/onboarding/ui'
 
-export function OnboardingMorePage() {
+export function OnboardingMore() {
   return <StepMore />
 }
 
-export default OnboardingMorePage
+export default OnboardingMore
