+++ libs/pages/onboarding/feature/src/lib/components/onboarding-personalize/onboarding-personalize.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react'
+
+import OnboardingPersonalize from './onboarding-personalize'
+
+describe('OnboardingPersonalize', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<OnboardingPersonalize />)
+    expect(baseElement).toBeTruthy()
+  })
+})
