+++ libs/pages/onboarding/feature/src/lib/components/onboarding-pricing/onboarding-pricing.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react'
+
+import OnboardingPricing from './onboarding-pricing'
+
+describe('OnboardingPricing', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<OnboardingPricing />)
+    expect(baseElement).toBeTruthy()
+  })
+})
