+++ libs/pages/onboarding/feature/src/lib/components/onboarding-project/onboarding-project.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react'
+
+import OnboardingProject from './onboarding-project'
+
+describe('OnboardingProject', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<OnboardingProject />)
+    expect(baseElement).toBeTruthy()
+  })
+})
