+++ libs/pages/onboarding/feature/src/lib/components/onboarding-project/onboarding-project.tsx
@@ -1,0 +1,10 @@
+import { StepProject } from '@console/pages/onboarding/ui'
+import { useDocumentTitle } from '@console/shared/utils'
+
+export function OnboardingProject() {
+  useDocumentTitle('Onboarding Project - Qovery')
+
+  return <StepProject />
+}
+
+export default OnboardingProject
