+++ libs/pages/onboarding/feature/src/lib/onboarding.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react'
+
+import Onboarding from './onboarding'
+
+describe('Onboarding', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<Onboarding />)
+    expect(baseElement).toBeTruthy()
+  })
+})
