+++ libs/pages/onboarding/feature/src/lib/onboarding.tsx
@@ -1,16 +1,16 @@
 import { Route, Routes } from 'react-router'
-import { LayoutOnboarding } from '@console/pages/onboarding/ui'
+import { Container } from './components/container/container'
 import { ROUTER_ONBOARDING } from './router/router'
 
 export function OnboardingPage() {
   return (
-    <LayoutOnboarding routes={ROUTER_ONBOARDING}>
+    <Container>
       <Routes>
         {ROUTER_ONBOARDING.map((route) => (
           <Route key={route.path} path={route.path} element={route.component} />
         ))}
       </Routes>
-    </LayoutOnboarding>
+    </Container>
   )
 }
