--- libs/pages/onboarding/feature/src/lib/pages/onboarding-company-page/onboarding-company-page.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react'
-
-import OnboardingCompanyPage from './onboarding-company-page'
-
-describe('OnboardingCompanyPage', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<OnboardingCompanyPage />)
-    expect(baseElement).toBeTruthy()
-  })
-})
