--- libs/pages/onboarding/feature/src/lib/pages/onboarding-company-page/onboarding-company-page.tsx
@@ -1,7 +1,0 @@
-import { StepCompany } from '@console/pages/onboarding/ui'
-
-export function OnboardingCompanyPage() {
-  return <StepCompany />
-}
-
-export default OnboardingCompanyPage
