--- libs/pages/onboarding/feature/src/lib/pages/onboarding-more-page/onboarding-more-page.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react'
-
-import OnboardingMorePage from './onboarding-more-page'
-
-describe('OnboardingMorePage', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<OnboardingMorePage />)
-    expect(baseElement).toBeTruthy()
-  })
-})
