--- libs/pages/onboarding/feature/src/lib/pages/onboarding-personalize-page/onboarding-personalize-page.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react'
-
-import OnboardingPersonalizePage from './onboarding-personalize-page'
-
-describe('OnboardingPersonalizePage', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<OnboardingPersonalizePage />)
-    expect(baseElement).toBeTruthy()
-  })
-})
