--- libs/pages/onboarding/feature/src/lib/pages/onboarding-personalize-page/onboarding-personalize-page.tsx
@@ -1,7 +1,0 @@
-import { StepPersonalize } from '@console/pages/onboarding/ui'
-
-export function OnboardingPersonalizePage() {
-  return <StepPersonalize />
-}
-
-export default OnboardingPersonalizePage
