--- libs/pages/onboarding/feature/src/lib/pages/onboarding-pricing-page/onboarding-pricing-page.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react'
-
-import OnboardingPricingPage from './onboarding-pricing-page'
-
-describe('OnboardingPricingPage', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<OnboardingPricingPage />)
-    expect(baseElement).toBeTruthy()
-  })
-})
