--- libs/pages/onboarding/feature/src/lib/pages/onboarding-pricing-page/onboarding-pricing-page.tsx
@@ -1,7 +1,0 @@
-import { StepPricing } from '@console/pages/onboarding/ui'
-
-export function OnboardingPricingPage() {
-  return <StepPricing />
-}
-
-export default OnboardingPricingPage
