--- libs/pages/onboarding/feature/src/lib/pages/onboarding-project-page/onboarding-project-page.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react'
-
-import OnboardingProjectPage from './onboarding-project-page'
-
-describe('OnboardingProjectPage', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<OnboardingProjectPage />)
-    expect(baseElement).toBeTruthy()
-  })
-})
