--- libs/pages/onboarding/feature/src/lib/pages/onboarding-project-page/onboarding-project-page.tsx
@@ -1,7 +1,0 @@
-import { StepProject } from '@console/pages/onboarding/ui'
-
-export function OnboardingProjectPage() {
-  return <StepProject />
-}
-
-export default OnboardingProjectPage
