+++ libs/pages/onboarding/feature/src/lib/router/router.tsx
@@ -6,31 +6,31 @@ import {
   ONBOARDING_PROJECT_URL,
   Route,
 } from '@console/shared/utils'
-import { OnboardingPersonalizePage } from '../pages/onboarding-personalize-page/onboarding-personalize-page'
-import { OnboardingMorePage } from '../pages/onboarding-more-page/onboarding-more-page'
-import { OnboardingCompanyPage } from '../pages/onboarding-company-page/onboarding-company-page'
-import { OnboardingPricingPage } from '../pages/onboarding-pricing-page/onboarding-pricing-page'
-import { OnboardingProjectPage } from '../pages/onboarding-project-page/onboarding-project-page'
+import { OnboardingPersonalize } from '../components/onboarding-personalize/onboarding-personalize'
+import { OnboardingMore } from '../components/onboarding-more/onboarding-more'
+import { OnboardingCompany } from '../components/onboarding-company/onboarding-company'
+import { OnboardingPricing } from '../components/onboarding-pricing/onboarding-pricing'
+import { OnboardingProject } from '../components/onboarding-project/onboarding-project'
 
 export const ROUTER_ONBOARDING: Route[] = [
   {
     path: ONBOARDING_PERSONALIZE_URL,
-    component: <OnboardingPersonalizePage />,
+    component: <OnboardingPersonalize />,
   },
   {
     path: ONBOARDING_COMPANY_URL,
-    component: <OnboardingCompanyPage />,
+    component: <OnboardingCompany />,
   },
   {
     path: ONBOARDING_MORE_URL,
-    component: <OnboardingMorePage />,
+    component: <OnboardingMore />,
   },
   {
     path: ONBOARDING_PRICING_URL,
-    component: <OnboardingPricingPage />,
+    component: <OnboardingPricing />,
   },
   {
     path: ONBOARDING_PROJECT_URL,
-    component: <OnboardingProjectPage />,
+    component: <OnboardingProject />,
   },
 ]
