+++ libs/pages/onboarding/ui/src/index.ts
@@ -3,4 +3,5 @@ export * from './lib/components/step-pricing/step-pricing'
 export * from './lib/components/step-more/step-more'
 export * from './lib/components/step-personalize/step-personalize'
 export * from './lib/components/step-company/step-company'
+export * from './lib/components/benefits-card/benefits-card'
 export * from './lib/components/layout-onboarding/layout-onboarding'
