+++ libs/pages/overview/feature/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/overview-page'
+export * from './lib/overview'
