+++ libs/shared/ui/src/index.ts
@@ -2,7 +2,9 @@
 export * from './lib/enums/icon.enum'
 export * from './lib/enums/plan.enum'
 // types
-export * from './lib/types/value.interface'
+export * from './lib/types/common/value.interface'
+export * from './lib/types/onboarding/price.interface'
+export * from './lib/types/onboarding/plan.interface'
 // components
 export * from './lib/components/icon/icon'
 export * from './lib/components/inputs/input-select/input-select'
