+++ libs/shared/ui/src/lib/components/inputs/input-text/input-text.tsx
@@ -55,6 +55,7 @@ export function InputText(props: InputTextProps) {
         <input
           id={label}
           name={name}
+          className="input__value"
           type={type}
           value={value}
           ref={inputRef}
