+++ libs/shared/ui/src/lib/types/common/value.interface.ts
