+++ libs/shared/ui/src/lib/types/onboarding/plan.interface.ts
@@ -1,0 +1,11 @@
+import { Value } from '../common/value.interface'
+import { Price } from '../onboarding/price.interface'
+
+export interface Plan {
+  name: string
+  title: string
+  text: string
+  price?: number
+  listPrice: Price[]
+  listDeploy: Value[]
+}
