+++ libs/shared/ui/src/lib/types/onboarding/price.interface.ts
@@ -1,0 +1,4 @@
+export interface Price {
+  number: string
+  price: string
+}
